Fix handling of protos in dotted package
When a proto file with `package foo.bar;` is parsed, the initial proto
object is nested like `{ foo: { bar: { <proto object> } } }`.
The code failed to handle this case, accessing the structure as
`proto['foo.bar']` rather than the correct `proto.foo.bar`.
This change treats the `serverNode.protoPackage` as a path, rather than
a key, insto the proto object. We use lodash's `get` function, which
takes care of resolving the path.